+++ test_fuc.py
@@ -1,50 +1,55 @@
 import unittest
+import subprocess
 
 from fuc.api.common import fuc_dir
-from fuc.api.VcfFrame import VcfFrame
-from fuc.api.BedFrame import BedFrame
-from fuc.api.FastqFrame import FastqFrame
+from fuc import pyvcf, pybed, pyfq
 
-class TestVcfFrame(unittest.TestCase):
+class TestPyvcf(unittest.TestCase):
 
     def test_filter_bed(self):
-        vf = VcfFrame.from_file(f'{fuc_dir()}/data/vcf/1.vcf')
-        bf = BedFrame.from_file(f'{fuc_dir()}/data/bed/1.bed')
+        vf = pyvcf.read_file(f'{fuc_dir()}/data/vcf/1.vcf')
+        bf = pybed.read_file(f'{fuc_dir()}/data/bed/1.bed')
         vf = vf.filter_bed(bf)
         self.assertEqual(vf.df.shape, (3, 13))
 
     def test_merge(self):
-        vf1 = VcfFrame.from_file(f'{fuc_dir()}/data/vcf/1.vcf')
-        vf2 = VcfFrame.from_file(f'{fuc_dir()}/data/vcf/2.vcf')
+        vf1 = pyvcf.read_file(f'{fuc_dir()}/data/vcf/1.vcf')
+        vf2 = pyvcf.read_file(f'{fuc_dir()}/data/vcf/2.vcf')
         vf3 = vf1.merge(vf2, how='outer', format='GT:DP')
         self.assertEqual(vf3.df.shape, (7, 14))
 
     def test_compare(self):
-        vf = VcfFrame.from_file(f'{fuc_dir()}/data/vcf/1.vcf')
+        vf = pyvcf.read_file(f'{fuc_dir()}/data/vcf/1.vcf')
         self.assertEqual(vf.compare('Steven', 'Sarah'), (0, 1, 1, 3))
 
     def test_filter_multiallelic(self):
-        vf = VcfFrame.from_file(f'{fuc_dir()}/data/vcf/1.vcf')
+        vf = pyvcf.read_file(f'{fuc_dir()}/data/vcf/1.vcf')
         vf = vf.filter_multiallelic()
 
     def test_reset_samples(self):
-        vf = VcfFrame.from_file(f'{fuc_dir()}/data/vcf/1.vcf')
+        vf = pyvcf.read_file(f'{fuc_dir()}/data/vcf/1.vcf')
         vf = vf.reset_samples(['Sarah', 'John'])
 
-class TestBedFrame(unittest.TestCase):
+class TestPybed(unittest.TestCase):
 
     def test_intersect(self):
-        bf1 = BedFrame.from_file(f'{fuc_dir()}/data/bed/1.bed')
-        bf2 = BedFrame.from_file(f'{fuc_dir()}/data/bed/2.bed')
-        bf3 = BedFrame.from_file(f'{fuc_dir()}/data/bed/3.bed')
+        bf1 = pybed.read_file(f'{fuc_dir()}/data/bed/1.bed')
+        bf2 = pybed.read_file(f'{fuc_dir()}/data/bed/2.bed')
+        bf3 = pybed.read_file(f'{fuc_dir()}/data/bed/3.bed')
         bf4 = bf1.intersect(bf2)
         self.assertEqual(bf3.to_string(), bf4.to_string())
 
-class TestFastqFrame(unittest.TestCase):
+class TestPyfq(unittest.TestCase):
 
     def test_shape(self):
-        qf = FastqFrame.from_file(f'{fuc_dir()}/data/fastq/1.fastq')
-        self.assertEqual(qf.shape, 4)
+        qf = pyfq.read_file(f'{fuc_dir()}/data/fq/1.fastq')
+        self.assertEqual(qf.df.shape, (5, 4))
+
+class TestCli(unittest.TestCase):
+
+    def test_qfcount(self):
+        result = subprocess.run(['fuc', 'qfcount', f'{fuc_dir()}/data/fq/1.fastq'], capture_output=True, text=True, check=True)
+        self.assertEqual(int(result.stdout.strip()), 5)
 
 if __name__ == '__main__':
     unittest.main()