+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * :issue:`2`: Fix Read the Docs automodule not working properly.
 * :issue:`3`: Add ``sphinx-issues`` extension to Read the Docs.
 * Add the ``VcfFrame.combine`` and ``VcfFrame.subtract`` methods.
+* Rename the ``BedFrame``, ``FastqFrame``, and ``VcfFrame`` modules to ``pybed``, ``pyfq``, and ``pyvcf``, respectively
 
 0.1.4 (2021-04-21)
 ------------------